Configure the IntermittentUnitBlock to avoid changes that cannot be handled
The MPSolver cannot currently deal with certain modifications (in particular,
a variable bound that is initially zero and becomes nonzero). Therefore, when
using BundleSolver to solve the dual of the UCBlock, we prevent these changes
from happening by considering a small positive value in the place of zero in
certain bound constraints.